refactor : 패키지명 participant -> player 로 변경
