+++ src/main/java/domain/player/Participant.java
@@ -1,4 +1,4 @@
-package domain.participant;
+package domain.player;
 
 import domain.area.CardArea;
