+++ src/main/java/domain/player/State.java
@@ -1,4 +1,4 @@
-package domain.participant;
+package domain.player;
 
 public enum State {
     HIT,
