+++ src/test/java/domain/player/ParticipantTest.java
@@ -1,4 +1,4 @@
-package domain.participant;
+package domain.player;
 
 import domain.area.CardArea;
 import domain.card.Card;
