+++ src/test/java/domain/player/StateTest.java
@@ -1,7 +1,6 @@
-package domain.participant;
+package domain.player;
 
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
